Start keeping TPA release notes
These release notes are meant to be read by 2ndQuadrant engineers, and
their purpose is to help people catch up quickly with recent changes.

I will update this file when I merge in a new feature or bugfix branch,
and remove/edit entries when tagging the release. (So there will be more
detail for the unreleased current code while we work on it, but once we
tag a release, we can discard less significant items.)