Move utils/ to bin/, retain symlink
Now it's possible to put $TPA_DIR/bin in your PATH, and just run
"provision cluster_name" and so on. As before, the cluster_name may be
either an absolute path or a directory relative to $TPA_DIR/clusters.

The ansible wrappers (ansible{,-doc,-vault,-playbook} move into ansible/
because we don't want them in $PATH… but just as a matter of caution, we
also make ANSIBLE_HOME override PATH in ansible/ansible; if it's set, we
always use it even if there's an ansible binary in our PATH.

We also stop trying to use paths relative to TPA_DIR, e.g., in ssh file
arguments in the inventory. We say «-o UserKnownHostsFile=known_hosts»
and chdir into the cluster directory before running playbooks.