Add support for cascading replication setup
We need to run standby clone with --upstream-conninfo pointing to the
upstream we want to follow (but not the one we want to clone from,
because it doesn't exist yet), and we need to run standby register
with --upstream-node-id.

We can't specify --upstream-node-id to standby clone, because then
repmgr will try to connect to that node and fail when it finds that the
node doesn't exist yet.