Make instance termination conditional on groups[cluster_tag]
Earlier, we used to fail if groups[cluster_tag] was not defined, i.e.,
if the instances in the cluster had already been terminated. This was
unnecessary and counterproductive because it meant we couldn't remove
any remaining resources (subnets, security groups, VPCs) if the EC2
instances had already been terminated.

Now we terminate the instances if they're there (according to the
inventory), but if not, we'll continue on to other resources.