Be conservative about which keypairs we remove
We can't tag them with CreatingCluster, and it's possible to use
existing keypairs, so we don't try to remove *any* keypair that was used
by an instance; instead, we only delete tpa_cluster_name, because we can
be reasonably confident that a key thus named is one we generated.

So we now remove keypair, instance profile, and policy by name; and use
CreatingCluster for EBS volumes, security groups, subnets, and VPCs.