Separate igw creation from VPC creation
In most cases, config.yml should set «ec2_instance_reachability: public»
to reproduce the earlier behaviour (of creating an igw in every VPC and
making all the subnets public—though the latter is not implemented in
this commit).

Otherwise, ec2_vpc_igw can be set (see the comments in provision.yml) to
control which regions get an igw and which not.