Separate route table creation from subnet creation
The original motivation for this change was that we were creating some
subnets and failing without creating routing tables on one run, and then
subsequently using the subnets created earlier without ever creating a
routing table, rendering instances unreachable after provisioning.

Separating these two tasks makes it possible to support more complex
routing configurations anyway. This is the minimum change possible in
this area: we just do what we did before, but in a different place. We
support only a single route to the igw, and there's no per-region or
per-subnet configurability yet.