Make vars.json opt-in
You must run «bin/provision … -e use_cached_vars=yes» to use vars.json
(and the earlier outdated-cache checks still apply, so it'll recompute
provisioning vars if it can tell something has changed).

This is a partial concession to reports that bin/provision has failed to
pick up changes to config.yml (especially ec2_ami). I have not been able
to reproduce the problem(s), and they may have been fixed, but vars.json
was always meant as an aid to development, so it's appropriate to hide
the cache behind a switch.