Overhaul vars.json interactions
The purpose of vars.json is to store provisioning information that is
expensive to recompute, in order to avoid extra work on subsequent runs.

It was originally added to speed up provisioning cycles in development
(and subsequently disabled by default). As a convenience, it tries to
detect config.yml changes that could potentially invalidate the cache,
but as a "best-effort" thing, not meant to be comprehensive.

This commit does multiple things:

1. Clean up the use of vars.json in various minor ways, add comments
   about the conditions for which the cache is invalidated, write the
   logic more clearly, and so on.

2. Extend the caching to more slow operations: VPC and subnet searches,
   routing table creation, and volume tagging. (This halves the time to
   rerun provision on my machine.)

3. Add set-provisioning-vars.yml, which allows multiple variables to be
   set in vars.json at once, avoiding multiple (slow) writes.

4. [As an aside] Iterate over instances/instance_vars/etc. and build
   custom data structures that are then used to invoke modules like
   ec2_vpc_subnet and ec2_tag in a loop, to make the output from the
   latter more palatable (and the data extraction more debuggable).