Simplify inventory cleanup during deprovisioning
It turns out that the file module will remove non-empty directories if
you say state=absent, even if both recurse/force are false. So the old
code would remove the inventory directory even if there were files in
it that were not generated. Oops.

Also, there's no point being POSIX-ly correct and avoiding the use of
«-mindepth 1» and then going on to use «-delete» (which is also a GNU
extension).

Anyway, what we end up with is an only slightly-awkward way to write

    find dir -type d -empty -delete

Portability is painful.