Hoist cluster_facts from "host fact" to "set_fact fact"
Ansible recently changed the precedence of facts returned by modules to
be lower than facts set with set_fact (this is how the precedence was
documented to work, but didn't), so rerunning cluster_discovery would
not override the earlier (possibly empty) fact.

Now we always "hoist" the host fact to set_fact level, so that the code
operates on the latest cluster_facts, as desired.