Separate instances into aws_instances, bare_instances, etc.
The idea is that we start out with instances[], fill in the platform,
derive xxx_instances[] for each platform xxx, let the platform-specific
code (both validation and provisioning) operate on it without having to
worry about filtering out entries for other platforms.

Then at the end, we recombine the various provisioning results into
instance_vars and write it out to the common inventory.