Rename regions to aws_regions
While we're at it, we also validate that they're valid AWS region names.
(We should actually be trying to check against the list of regions known
to the installed boto, but that's more fiddly.)

The general rule of thumb is that platforms/xxx/validate.yml may define
various xxx_* facts for use by platforms/xxx/provision.yml; at the very
least, other platforms are discouraged from using such a fact.