Allow instance defaults to be specified at the top level
'platform' is not the only thing that we may want to declare at the top
level and override for individual instances. Rather than treat it as a
special case, we follow the expand_instance_tags→set_instance_defaults
name change to its logical conclusion and allow a default value to be
specified for any instance key.

So now you can do something like this

    instance_defaults:
      region: eu-west-1
      type: t2.micro

    instances:
      - node: 1
        Name: foo
        …

      - node: 2
        Name: bar
        …

      - node: 3
        Name: baz
        region: eu-west-2
        type: t2.medium

and the instances will have the correct type and region set by the time
they are validated.

In large homogeneous clusters, this can make the definition more concise
(especially since you can even hoist identical volumes[] definitions to
the defaults).