Add minimal support for platform 'bare'
This is meant for existing servers to which we are given access, where
we don't need to perform any provisioning tasks, just add them to the
inventory with definitions from config.yml.

Note that we currently handle ssh user/hostkeys wrong for such servers
(e.g., we will add entries for them to known_hosts without ever setting
their hostkeys).