Simplify AMI-name-based if-soup in user-data
Now we can work harder to identify a distribution from the name of an
AMI in one place, and just use the result anywhere else that needs it.
Also other minor cleanups.

It might be nice if we had a (TPA,version) pair also available in
user-data, so that we can test for the TPA version rather than just
assume all TPA images will have whatever we're testing for (e.g., if
mdadm was added in TPA-*-xxx images).

XXX Note that if we need to install both awscli and mdadm for some
reason on Debian, we'll run apt-get update twice. It's probably not
worth trying to avoid it, because it doesn't happen often.