Use include_tasks for platforms/*/{,de}provision.yml
I wanted to use import_tasks so that "tags: xxx" on the import in
platforms/provision.yml would push the tags down to all imported tasks,
but there's a downside, in that tasks that evaluate variables resulting
from skipped tasks might fail.

We stick with import_tasks for things tagged "always", like validation,
and switch to dynamic include_tasks for provisioning and deprovisioning.
That means our tags are a mess and will need to be redone, but that's
not really something new.