Make sure ec2.py succeeded before deployment
In the past few days, I have been seeing frequent timeouts from ec2.py,
after which deploy continues blithely onwards with no ec2_xxx variables
defined. That doesn't work very well, and it's not always obvious why it
broke. So we make sure that ec2_id is defined in roles/platforms/aws.

As part of this change, we write platform=xxx to the static inventory.