Handle IP addresses for bare instances
Bare instances must specify at least one IP address explicitly in the
configuration (unlike platforms/aws, where we can discover the current
values for the static inventory, and ec2.py will later augment that),
so we require that and pass the information onwards to the inventory.

Aside: we also set instances[*].vars to {} by default.