Remove references to 2ndQuadrant/ansible
Instead, we'll use the upstream stable-2.6 branch going forward (2.6 has
not yet been released, but the branch exists, and it should be released
soon). The 2ndQuadrant/ansible repository still exists, but the master
branch has been updated to point to remotes/upstream/stable-2.6.

Most of the commits in 2ndQuadrant/ansible have either been committed
upstream over the years, or are not critical; using upstream 2.6 means
that we can stop having to answer questions about 2ndQuadrant/ansible.