Start to revise repository/package handling
We want to install all required repositories in one place, so that the
configuration is easy to manage, and also so that handling clusters with
no outside network access can be done in one place. This is only a first
step towards that change, but it's a tedious and complicated one.

Before we install packages, we need to bootstrap Python so that we can
run Ansible modules (before that, we can only run «raw» tasks). First we
install Python 2.7.x, and we also need to install the ssl_match_hostname
backport for Python <2.7.9.

We used to install the backport using pip. To get pip, we would install
python-pip. To get python-pip (on RedHat), we would install EPEL. But we
can avoid that if we install python-virtualenv from the main repository,
because creating a virtualenv installs pip therein, even though it's not
in any system package. After I figured that out, I also figured out that
the backport is available in system packages, and we can just use those.

That leaves us with awscli, which we can install into the virtualenv. It
does mean it won't be available globally, but since we currently don't
use it during deployment, that doesn't matter too much: future scripts
that use it can easily activate the virtualenv first.

The one place we already need awscli is in platforms/aws/user-data,
where we can also avoid installing EPEL and use pip via virtualenv.
(I tested that «virtualenv x && pip install y» in user-data is still
compatible with «pip: name=y virtualenv=x» in deployment playbooks.)

Because we don't need EPEL or pip to bootstrap Python, we can move the
awscli installation to the end and use a «pip» module invocation instead
of a «raw» task as before. And we also switch to installing the package
for python-psycopg2 in postgres/src (we can revisit any problems that
causes later).

Finally, we introduce a new OS-specific repositories.yml stage where we
can configure any repositories we like. It's empty at the moment.