Introduce a default/override mechanism for packages
At this stage, we (will) know we have all the required repositories
enabled, so we can allow the caller to specify any packages they may
want, and install them all in the same transaction.

Note that the default package lists can be overriden only by overriding
default_xxx. That's inconvenient, but it's OK, because you shouldn't do
it in the first place.