Overhaul repository configuration
By default, we'll install PGDG, 2Q (public|internal|testing|snapshot),
and (when available) 2ndQPostgres repositories (both apt and yum), but
we allow other repositories to be specified (though this should not be
needed in normal situations).

We don't try to do any pinning or exclusions, depending instead on the
package versions to define the right priority for each repository.