Set inventory.any_unparsed_is_failed in ansible.cfg
This fixes the problem in deprovision where ec2.py fails and we try to
deprovision stuff without removing the instances, and it fails because
the volumes we're trying to remove are still attached to instances we
can't see.

(Provision had a similar problem, but the platform=='aws'&&ec2_id check
solves the problem in that case.)

The code to act on this setting was merged into upstream devel via
https://github.com/ansible/ansible/pull/41171, but is currently only
available in 2.7-to-be. It may be considered for backporting into
stable-2.6, we'll have to wait and see.