Revise roles/platforms and move ip_addresses to common/hostvars
First, we get rid of the meta dependencies in favour of include_role
(which allows us to do things in a more sensible order, such as check
that we're running Ansible 2.6.0 before we check that EC2 hosts are OK.)

Second, we give individual platforms an opportunity to set
public_ip/private_ip and then set ip_addresses[] in common.

Plus miscellaneous shuffling (e.g., platforms/vars was pointless).