Allow instance settings to be added to vars without duplication
Certain settings in instances[] may also be needed in the inventory vars
for use in templates and so on. This change allows them to be copied to
vars without duplicating the value in the definition.

For example:

    - node: 1
      …
      route53_hosted_zone_id: xxx
      route53_zone: yyy
      xxx: 42
      export_as_vars:
        - route53_hosted_zone_id
        - route53_zone
        - xxx
      vars:
        …
        xxx: 53

This would add route53_zone and route53_hosted_zone_id and xxx to the
inventory for the given host.

Note that export_as_vars can also be specified in instance_defaults[],
and also that setting the same key explicitly in vars will override the
exported setting (xxx above will be 53, not 42).