Make bin/* accept only an absolute or relative path to the cluster
We no longer interpret "foo" as $TPA_DIR/clusters/foo but as "./foo".
You can still say "~/path/to/foo" if you want.

This is important because once TPA is installed from packages, people
will be creating separate cluster directories, not using the bundled
clusters (which won't even be bundled anyway).

Note that cd $cluster will still fail with an error message if the given
directory does not exist; that's why I got rid of the -d test.