Add 'architectures/Training'
This supersedes clusters/training (which now goes away) and brings
"generate-config.sh" under the tpaexec config framework.

This is an example of how architectures/*/configure can do its own
option parsing and then generate config.yml. In this case, it does so
with a shell script and :cat:, but it could have been a shell script
that called «ansible -m template», or a Perl program, or whatever.

Note that it's possible to run the configure script by hand if you pass
it all the arguments it expects (the path to an existing directory and
then whatever other parameters it expects), but obviously we expect it
to be run through tpaexec.

See 'tpaexec info architectures/Training' for more details.