Move the «mkdir $cluster» from tpaexec to configure
In the spirit of letting the architecture's configure script have full
control over the configuration process, we don't even create (or check
for) the target cluster directory in tpaexec any more.

Most configure scripts can just «mkdir $cluster» and leave it to set -e
to fail if the target already exists. But they could choose to support
reconfiguration of an existing cluster and do something different if
they wanted to.