Move some boilerplate and common option parsing to lib
configure scripts can now handle the options they care about and source
the common options file to handle the rest.

This means that we can document common options in one place, and refer
to it from elsewhere.

(This also means that configure scripts written in other languages will
have to make sure to duplicate the handling of the common options, which
should not be much of a problem.)