Add minimal_setup module for distribution discovery
We used to set ansible_distribution{,_release}, ansible_os_family, and
ansible_pkg_mgr with some raw tasks and set_fact because we weren't sure
yet if a usable Python was installed on the target (so the setup module
might fail, and even if it worked it was too slow at that time, before
gather_subset=min was introduced).

We used the results both in conditions and to make the package module
work (based on ansible_pkg_mgr). In 2.6, with the namespacing of host
facts and their separation from vars, package no longer looks at the
ansible_pkg_mgr var, as set by set_fact. You also can't directly set
ansible_facts.pkg_mgr, though «set_fact: pkg_mgr=… cacheable=false»
might have been close enough.

But at the same time, the precarious tower of raw tasks was hard to
extend to detect different distributions. It would have been nice to
simplify matters and reduce roundtrips by using the setup module, but
that still requires Python to be available on the target.

So we use a shell module that returns the correct host facts, which are
injected as vars by the [defaults]inject_facts_as_vars setting. We can
easily extend the module without adding roundtrips, and the package
module doesn't need to perform another round of fact discovery.