Deprecate ec2_ami_user and cluster_ssh_user
The situation with ec2_ami_user, cluster_ssh_user, and ansible_user is…
unfortunate. All of them mean something slightly different, and the one
that's most meaningful is the one we use the least. Before the confusion
proliferates further (e.g., the documentation suggests setting both), we
make ansible_user as an instance var the preferred form:

    instance_defaults:
      vars:
        ansible_user: admin

We continue to accept ec2_ami_user and cluster_ssh_user as top-level
config.yml settings and transform them into instance vars to maintain
compatibility. If vars.ansible_user is set in either instance_defaults
or instances, it will override the top-level setting (like any other
instance var).

We stop writing ansible_user as a group var and instead write it as a
host var (like any other instance var). We also write 'User xxx' to the
individual entries in ssh_config, and not to the 'Host *' defaults. With
these changes we use an instance variable for an instance-scoped setting
without trying to shoehorn the functionality into a global setting.

(The above argument applies to cluster_ssh_port, but in a weaker form
because there aren't multiple competing settings, and it is less likely
for there to be a good reason to use different ssh ports, and there's an
easy workaround (i.e., setting vars.ansible_port) available.)

Miscellaneous notes:
  (a) ec2_ami_user made it hard to support hybrid clusters that used
      different AMIs (one global setting declared one way, and then
      per-instance settings declared in an entirely different way).
  (b) cluster_ssh_user was just a lame generalisation of ec2_ami_user
  (c) cluster_vars.cluster_ssh_user never worked
  (d) in theory you could set cluster_vars.ansible_user and have it work
      like anything else in group_vars, but there's no compelling reason
      to recommend that.

Documentation update pending review of this commit.