Simplify --hostnames-xxx option passing
Rather than pollute every invocation of lib/hostnames with a ridiculous
workaround for an ancient bash bug, we can set environment variables in
configure-options and make lib/hostnames accept them.

While we're at it, we also allow the sort order to be specified by the
same mechanism, so you can now do something like this:

    tpaexec config … --hostnames-sorted-by --dictionary-order
    tpaexec config … --hostnames-sorted-by "--dictionary-order -r"

…where the extra argument is passed through unmodified to sort(1). This
is at least a little evil, so I've left it undocumented for now, but the
behaviour is actually useful because it allows for repeatable naming (so
for example, one can decide to have predictable names assigned to nodes
in a BDR cluster).