Add a modified copy of ec2.py to $TPA_DIR/library
We've had a long-standing problem where instance creation could fail
with an error like this:

    "Instance creation failed => InvalidParameterValue: Value
    (xxx_profile) for parameter iamInstanceProfile.name is invalid.
    Invalid IAM Instance Profile name"

Rerunning provision would solve the problem.

This happens because the IAM instance profile we created has not yet
propagated fully by the time we request instances that use that profile.
It happens nearly every time on fast networks.

We modify ec2.py to use an "EC2Retry" wrapper that inherits from
CloudRetry and wraps run_instances, detects the InvalidParameterValue
exception with the 'iamInstanceProfile.name is invalid' message, and
retries that request a few times.

Incidentally, terraform has the same problem and used the same fix,
right up to using string matching on the error message to figure out
whether it's the right error or not:

https://github.com/terraform-providers/terraform-provider-aws/pull/3055