Make TPA_DIR detection more helpful
1. If TPA_DIR is explicitly set, we use it
2. If TPA_DIR is not set, and
    (a) We can infer a valid TPA_DIR from $0, we use that, or
    (b) We set TPA_DIR to /opt/2ndQuadrant/TPA

No matter which of the above cases is true, we check for $TPA_DIR/roles
at the end and complain if it doesn't exist.

Now (1) a packaged installation works whether /opt/2ndQuadrant/TPA/bin
is in your PATH or you copy tpaexec into ~/bin, (2) if you run tpaexec
from a git checkout, it takes precedence over the package install, and
(3) if you set TPA_DIR explicitly, we use that value.