Remove pycrypto, add cryptography
Marco pointed out that pycrypto was defunct (though it's still supported
by default in Ansible), but Ansible supports cryptography, which is easy
to install and doesn't need a compiler either.

Also remove version limits on various modules (ansible doesn't specify
any, and the version problems with jinja2 are a thing of the past).