Don't set/use cluster_network{,s} by default
First, if you have cluster_network{,s} set in config.yml, everything
should continue to work as before: cluster_network will be converted to
cluster_networks (if the latter is not specified), and pg_hba.conf will
permit connections from anything in cluster_networks.

But we no longer specify cluster_network in config.yml by default, nor
set the default value of all subnets if it's not specified, nor require
it to be set, nor override it when configuring openvpn. Instead, we add
sets of per-host entries to pg_hba.conf based on their role.