Reorganise roles/test a little, record pg_controldata output
We now define a timestamped output_dir in lib/test.yml, depend on its
being defined in roles/test, and keep timestamps out of the filename, so
that it's easy to find all the results of a single test run in the same
place.

We also define an include-able helper to write output to
{{output_dir}}/{{output_file}}.

Also remove test output when deprovisioning the cluster (is that really
the right thing to do? I can't quite decide).