Introduce new unconditional 'final' role
Instead of doing

    - role: postgres/final
      when: >
        'postgres' in role

we now apply the new role "final" to all instances. Now postgres/final
doesn't have to worry about calling {pgbouncer,haproxy}/final (both of
which we may want to apply to instances without 'postgres' in role).

In general, deploy.yml should invoke roles like pgbouncer and haproxy
with the appropriate conditions after postgres and before final, and
leave it to final to invoke the appropriate /final roles.

For clusters generated with tpaexec configure, of course, these changes
will take effect already because of the changes to the architecture's
own deploy.yml.

For hand-written cluster configurations, this is an important change.
Without it, replication/backups will not work correctly.