Make haproxy support work
If you set haproxy_backends in vars on an instance, haproxy will now be
configured to use those backends.

XXX Right now we have to add a 'host' line to pg_hba.conf to allow
haproxy connections. We should figure out what to do to make it use
hostssl when appropriate.