Make sure the pgbouncer user is created in postgres
The pgbouncer role invokes postgres/createuser, but if postgres is not
running on the same instance, then postgres/final will not be invoked
later to create the user in Postgres.

Rather than try to delegate postgres/createuser to the right postgres
instance or instances (in the haproxy_backends case), we now apply the
pgbouncer role unconditionally, and invoke postgres/createuser on both
postgres and pgbouncer instances, and the rest on pgbouncer instances.

Not a very refined solution, but a simple one.

Fixes #210