Use destination_format/destination_format_tags to set ansible_host
Earlier, we used to set vpc_destination_variable to private_ip_address
or ip_address (and destination_variable to {private,public}_dns_name,
but that was unused because we always create instances in a VPC).

Unfortunately, if you stop an instance, it no longer has an ip_address
assigned (though it still has a private_ip_address), and ec2.py would
leave it out of the inventory because vpc_destination_variable wasn't
set, so stopped instances just didn't show up in the inventory at all.
As a consequence, we would create duplicate instances if you had any
stopped instances while re-running provision.

To solve this, we use destination_format to set ansible_host, and set
destination_format_tags to private_ip_address or ip_address, as needed.
If it is set to ip_address and there is no ip_address, then ansible_host
ends up being set to "None", but the stopped instance still shows up in
the inventory.