Make 'instance_with_backup_of' prefer instances in the same region
When setting restore_command, it makes sense to use the "closest" barman
instance rather than the first one. We pass inventory_hostname in to the
instance_with_backup_of filter, so that it can look for a backed-up
instance in the same region. If it doesn't find one, it returns the
first backed-up instance as it used to do before.