Always assign hostnames the same way
We use "zero" as a dummy value for hostnames[0] so that instances[] can
use ${hostnames[$i]} for node $i (instead of hostnames[$i-1]), just for
clarity. It's clearer to do it this way rather than ask lib/hostnames
for an extra hostname and then throw away the first one, and worse
still, to do so in some places but not others.