Add a lightweight template processing script
The following two commands are roughly equivalent:

  template.py template.j2 vars.yml \
    $libdir/templates … > config.yml

  ansible localhost -m template \
    -a "src=template.j2 dest=config.yml" -e @vars.yml

Running ansible for a single template tasks takes on the order of 3s to
complete, and provides no easy way to make template includes from
$libdir/templates work (since a FileSystemLoader is initialised with the
cwd as its only root directory).

It's easier and (much) faster to use jinja2 directly, but that means we
don't have access to the usual Ansible-defined filters and tests, so we
use Ansible's Templar class instead.