Make lib/cluster-{vars,tags} produce single-line hash output
This commit solves two problems. First, with the earlier code, we could
have ended up with invalid output like this:

    cluster_tags: {}
      Stuff: added
      by: template

Second, it's hard to stuff multiline output into vars[] for later use.

Now both scripts take key=value pairs on the command line, and produce
single-line output with all of the required settings. So the configure
script only has to do something like:

    cluster_vars=$($libdir/cluster-vars …)
    cat <<CONFIG
    cluster_vars: $cluster_vars
    …
    CONFIG

We don't go to much trouble to quote arguments properly, and rely
instead on YAML being able to sort things out. We'll need to change that
if we ever need to pass in more than a small handful of known variables.