Rewrite BDR-Always-ON/configure to use template.py
Now that we have all of tpaexec configure's operating vars[] written to
a vars-XXX.yml file, we can use template.py (instead of a here-doc) to
generate config.yml for our new cluster.

This scheme is opt-in, i.e., existing configure scripts will remain
unaffected and continue to work exactly as before. (In fact, they don't
even have to put anything into vars[], but doing so is better because it
makes a later template.py conversion much easier.)