Move hostnames/cluster_tags to configure-defaults
If we set $instances to the number of instances before sourcing
configure-defaults.sh, we can move hostname setting out of the
individual configure scripts.

"Empty" cluster_tags and cluster_vars can also move, on the basis that
they can always be overwritten with extra arguments later.

We could probably move subnet assignment if we tried, but it doesn't
seem worth it, especially when region assignment isn't worked out yet.