Allow explicit mountpoint to override translation
We need to be able to identify the postgres_data and barman_data volumes
reliably. The obvious way to do this is to require them to be annotated
with "volume_for: xxx_data". But that would have meant that we couldn't
change the mountpoint, because the volume_for would always have been
translated to the default.

Now we allow a volume to declare what it's for and also set a
non-standard mountpoint, e.g.:

    - device_name: /dev/xvdb
      …
      vars:
        volume_for: postgres_data
        mountpoint: /var/lib/postgresql

With this change, we can identify the volume's purpose regardless of its
mountpoint. If you set the mountpoint without setting volume_for, then
the volume will not be correctly identified, which may have negative
consequences in future.