Add a 'tpaexec setup' command
This command will (optionally) create and activate a virtualenv and
install all the required python packages into it.

XXX At the moment, it creates and activates the virtualenv itself, but
punts to misc/tpa-pip-install.sh to install the packages, which will not
always work (e.g., when --virtualenv is used to set an alternative
location).