Automatically activate $TPA_DIR/tpa-virtualenv (if it exists)
If you use the default location for a virtualenv, then there's no need
to activate it from .bashrc. We can just activate it automatically from
tpaexec, which reduces one more thing that the user needs to care about.
Of course, you can still specify a custom location if you want (in which
case you are responsible for activating it yourself).