Copy libselinux-python into virtualenv if required
I think it might be better to just create the virtualenv with
--system-site-packages so that the system's libselinux-python bindings
are picked up automatically, but I haven't verified that this fixes the
problem (whatever it is).

If we use --system-site-packages, we might be able to cut down on our
dependency list to some extent (for the non-critical modules).