Absorb misc/tpa-pip-install.sh into tpaexec
This still depends on finding $TPA_DIR/pip-packages/ansible (i.e., it
will not work if the -deps bundle is not installed, but you do have
network access), but it's a simple translation of what was in the
earlier script. We can omit some complications because we know we
created and activated a virtualenv already.

XXX It's not clear to me why we're upgrading pip itself. It can't be to
make the find-links stuff work, since we're using that in the upgrade
command itself. If it's just to silence the warning from pip, then it
might not be worth it. I didn't try to find an answer.