Coax former clusters/images roughly into shape as architectures/Images
The major pieces are laid out in the way they should be (well enough to
generate AMIs), and a big hammer is no longer the more appropriate tool
for the remainder of the job.

In particular, we need some way to configure which sets of repositories
and packages to install, and we can't really do that until we support
the new repository layout.