Don't ignore mkdir error in write_vars
Since the function apparently doesn't inherit the -e, mkdir would fail
if the directory existed, and then we would happily go on to overwrite
config.yml and so on. Oops.

We'll need to find a better solution someday when we can reconfigure a
cluster on the fly, but there's no danger of that happening soon.