Remove provision/deploy/deprovision/rehydrate links
It's better to stay consistent with interface changes to avoid confusing
customers, and the less we pollute the (PATH) namespace, the better. If
someone insists on having "bin/provision", they can create the link to
tpaexec and it will still work fine.

But from a documentation perspective, it's best to stick to the one
consistent way.