Add a bdr-pre-node-creation hook
If hooks/bdr-pre-node-creation.yml exists in the cluster directory,
we'll include it when we're about to create a BDR node.

For now, this is intended for use by TPAexec developers, and there are
no particular guarantees made to the included tasks (in terms of vars
defined and so on). If you need to use this, look at the surrounding
tasks in roles/postgres/bdr/tasks/main.yml and adapt to the context.