Add support for installing twoq_repositories
If you do

    export TPA_2Q_SUBSCRIPTION_TOKEN=xxx

(where xxx is the per-customer token extracted from the /api/repository
URL that you get from the portal when you add a software subscription;
at the moment, there's no other way to get a token as far as I know)

and define a list of repositories to install

    twoq_repositories:
      - products/default/release
      - products/2ndqpostgres/snapshot
      - ci-spool/bdr3/snapshot

we will now download the corresponding registration scripts from the
portal's API using the given token and let it do whatever it wants.

Other than running the task only if the file corresponding to the
repository does not already exist in /etc/apt/sources.list.d or
/etc/yum.repos.d, we leave everything to the script.

Note that if you install many repositories, apt-get update will be run
many times. There's also no attempt to separate key installation from
repository installation (or, in fact, to install keys at all; we just
assume the script will do it).

Not all combinations of vhost/code/maturity_level make sense, and not
all are currently working (i.e., if you use them, the script will be
run, but apt-get update may fail afterwards).

dl/default/release is always installed, and there's no way to turn it
off. There's no way to use different tokens for different repositories
(which isn't a sensible thing to do with per-customer tokens, but does
depend on the tokens remaining unique per-customer forever).

There's no support for re-running the registration script if the token
changes (e.g., because all the credentials are rolled over after an
expiration period, or because one is rolled forward because of a
compromise or subscription expiration).