Don't create the pglogical extension by default
Having it installed and unused is supposed to be harmless, but it makes
it impossible for ordinary (i.e., unprivileged) users to drop tables or
other objects, because the hook that pglogical installs tries to check
if pglogical is active by reading pglogical.node, which fails because
the unprivileged user doesn't have access to the pglogical schema.

That's a (long-standing) bug that needs to be fixed, but until then we
don't need to install it by default unless it's actually needed.