Set application_name in --upstream-conninfo
If we let repmgr generate a conninfo for the upstream server, it will
include an application_name. repmgr node check later depends on this
being set to identify which entry in pg_stat_replication corresponds
to which replica.

If we pass in the upstream-conninfo, repmgr uses it unmodified, which
means that repmgr node check fails to identify the pg_stat_replication
entry (because the application_name shows up as walreceiver).

So we add in the application_name to the conninfo we provide repmgr.