Set hostname in aws/user-data
On the face of it, this is just a minor change where we set the hostname
a little earlier rather than during deployment. At most we might claim
it reduces the potential for confusion if you ssh to newly-provisioned
instances before running tpaexec deploy.

But there's a less obvious reason that this is a good idea. mdadm will
use the hostname as the "homehost" when labelling a RAID array that it
creates. If it does so before our hostname is set, it will later fail
when trying to reassemble the fragments of that array on a new instance
undergoing rehydration, whose cluster hostname is also not yet set, and
whose ip-x-y-z hostname is different from the original node's original
hostname. Setting the hostname at the beginning of user-data sidesteps
this problem.